requirements.txt
fastapi
uvicorn
crunchyroll.py
